Place Akamai Cache Clear Block
This commit introduces and applies a patch to the moderation_sidebar
contrib module that first verifies that the Akamai module is installed
and, if it is, then places the Akamai cache clear plugin block in the
moderation sidebar. I also added a one line SCSS rule `display: none` to
a div element that displayed the Refresh URL, which I thought was
unnecessary clutter visually.