MWES-2463: Akamai Block Styles
This places the Akamai Block in the topmost header region and adds some
CSS styles to make the presentation of the block more appealing and
appropriate given the surrounding design/context.

I was unable to build the CSS from the SCSS files locally/in Docker, so
I will have to review this closely once the styles build in the PR
environment or managed platform.