MWES-2760: Login Page Display SSO Block
This commit hides the default Drupal login form on /user/login with CSS
(display: none) and adds some text above the 'Login with your developer
account' button that explains to editors the new login process and who
to contact for assistance. It also adds a new button that will toggle
the visibility of the old, default Drupal login form in case we need to
login using the default Drupal form for some reason.

I was unable to get my assets to build with NPM locally, so I will have
to review this once it is deployed to a PR environment or managed
platform.